Add long-range chart history (up to 12mo) and default to indefinite retention
Range picker gains 30d/90d/6mo/12mo alongside the existing 1h/6h/1d/7d.
Chart-data endpoints (telemetry, radio-clients) now downsample via
TimescaleDB time_bucket() to a bounded ~600 points regardless of the
selected span - without this, a 12mo range at one row per ~30s would be
over a million rows per series, far too much to query or render. Uses
avg() for continuous metrics (rssi/noise/mcs/rate/retries/throughput) and
last() for discrete state (channel/bandwidth_mhz), since averaging a
channel number produces a meaningless fractional value. TelemetryPoint/
RadioClientPoint's rssi/noise/mcs fields move from int to float to match
the now-averaged values Pydantic actually receives.

TELEMETRY_RETENTION_DAYS now defaults to 0 (keep forever, no retention
policy applied) instead of 180 - db.py's ensure_retention_policies now
actively removes an existing policy when set to 0, not just skips adding
one, so this takes effect on restart without manual SQL.

+++ app/models.py
@@ -68,10 +68,15 @@ class DeviceStatus(BaseModel):
 
 
 class TelemetryPoint(BaseModel):
+    # rssi/noise/mcs are float here (unlike RadioSnapshot's int) because
+    # get_telemetry_history now averages them across a time_bucket - only
+    # channel/bandwidth_mhz stay int, since those use last() (the literal
+    # value at the end of the bucket) rather than an average, since they're
+    # discrete state, not a continuous metric.
     time: datetime
-    rssi: Optional[int] = None
-    noise: Optional[int] = None
-    mcs: Optional[int] = None
+    rssi: Optional[float] = None
+    noise: Optional[float] = None
+    mcs: Optional[float] = None
     rate_mbps: Optional[float] = None
     retries: Optional[float] = None
     channel: Optional[int] = None
@@ -80,10 +85,12 @@ class TelemetryPoint(BaseModel):
 
 
 class RadioClientPoint(BaseModel):
+    # rssi is float, not int, for the same reason as TelemetryPoint above -
+    # get_radio_client_history averages it across a time_bucket.
     time: datetime
     client_mac: str
     host: Optional[str] = None
-    rssi: Optional[int] = None
+    rssi: Optional[float] = None
     rate_mbps: Optional[float] = None
     retries: Optional[float] = None
